refactor: alterando o tipo do retorno

+++ Api2/Application/Interfaces/ICalcularJurosAsyncAppService.cs
@@ -4,6 +4,6 @@ namespace Api2.Application.Interfaces
 {
     public interface ICalcularJurosAsyncAppService
     {
-        Task<decimal> CalcularJurosAsync(ParametrosConsulta parametros);
+        Task<double> CalcularJurosAsync(ParametrosConsulta parametros);
     }
 }
